rudof_lib
Module rudof

Macros

iri

Structs

DCTAP

PrefixMap

Contains declarations of prefix maps which are used in TURTLE, SPARQL and ShEx

QueryShapeMap

QuerySolution

Represents one query solution

QuerySolutions

Represent a list of query solutions

RdfData

Generic abstraction that represents RDF Data which can be behind SPARQL endpoints or an in-memory graph or both The triples in RdfData are taken as the union of the triples of the endpoints and the in-memory graph

ResultShapeMap

Contains a map of the results obtained after applying ShEx validation

Rudof

This represents the public API to interact with rudof

SRDFSparql

Implements SRDF interface as a SPARQL endpoint

ShExFormatter

Struct that can be used to pretty print ShEx schemas

ShExSchema

ShExValidator

ShaclSchema

ShapeMapFormatter

Struct that can be used to pretty print Shapemaps

ShapeMapParser

ValidationReport

ValidatorConfig

This struct can be used to customize the behavour of ShEx validators

VarName

Enums

DCTAPFormat

DCTAP available formats

RDFFormat

Posible RDF formats

ReaderMode

Reader mode when parsing RDF data files

ShExFormat

Contains possible ShEx formats

ShaclFormat

SHACL Formats supported. Mostly RDF formats In the future, we could also support SHACL Compact format

ShaclValidationMode

Backend used for the validation.

ShapeMapFormat

Format of Shapemap files

UmlGenerationMode

ValidationStatus

Represents the current status of validation

Type Aliases

Result